Remove pointer indirection from ConnectionT::entryPointHandle
Change ConnectionT::entryPointHandle from CFLOBDDNodeHandle* (heap-allocated
pointer) to CFLOBDDNodeHandle (value). Eliminates per-Connection malloc/free
overhead and fixes a memory leak in operator=.

Break the circular include between connectionT.h and cflobdd_node.h by moving
#include "connectionT.h" to after CFLOBDDNodeHandle is fully defined. Fix
include order in 9 headers to ensure cflobdd_node.h precedes connectionT.h.

Also: fix PairProduct/TripleProduct Hash() (size seed + fmix64), wire up
PairProductMapBody::hashCheck, add isCanonical guard to all Canonicalize()
methods, reformat operator== in reduction_map.cpp and return_map_T.h.

+++ CFLOBDD/cross_product.h
@@ -90,7 +90,7 @@ class PairProductMapBody {//: public List<intpair> {
   bool operator==(const PairProductMapBody &p) const;
   intpair& operator[](unsigned int i);                       // Overloaded []
   unsigned int Size();
-  unsigned int hashCheck;
+  size_t hashCheck;
  public:
   bool isCanonical;              // Is this PairProductMapBody in *canonicalPairProductMapBodySet?
   static Hashset<PairProductMapBody> *canonicalPairProductMapBodySet;
